19_10_05 version

+++ pa1.c
@@ -69,11 +69,8 @@ static void set_timeout(unsigned int timeout)
  *   Return <0 on error
  */
 
-struct sigaction sa;
-void sigint_handler(int signo){
-	//kill(getpid(),SIGKILL);
-	alarm(__timeout);
-}
+
+
 int No_such=0;
 
 static int run_command(int nr_tokens, char *tokens[]) 
@@ -87,92 +84,47 @@ static int run_command(int nr_tokens, char *tokens[])
 	pid_t pidnum; // fork() 반환 값 받을 pid 변수
 	pid_t waitPid; // 자식 프로세스 종료 상태 받을 pid 변수
 
-
-	sigemptyset(&sa.sa_mask);
-	sa.sa_handler=sigint_handler;
-	sa.sa_flags=0; // 시그널 처리하는 동안 전달 시그널 블록 X
-
-	
-
-	// sigaction(), alarm() 이용
-	// cd, prompt, for, timeout 은 built in command 
-	// 나머지 명령어들은 external command 
-	
-
 	if(!strcmp(tokens[0],"cd")){
-
-		if(nr_tokens==1){
-			chdir(getenv("HOME"));
-			return 1; // 여기서 끝
-		}	
-		else if(!strcmp(tokens[1],"~"))	{
-			chdir(getenv("HOME"));
-			return 1; // 여기서 끝
-		}
-		else{
-			chdir(tokens[1]);
-			return 1; // 여기서 끝
-		}
+		if(nr_tokens==1)				{chdir(getenv("HOME"));	return 1;}
+		else if(!strcmp(tokens[1],"~"))	{chdir(getenv("HOME"));	return 1;}
+		else							{chdir(tokens[1]);	return 1; }
 	} // "cd" is built in, not external
 	
 	else if(!strcmp(tokens[0],"for")){
 		int forcount=atoi(tokens[1]); // count "for" command
-
-		for(int i=0;i<forcount;i++){
-			run_command(nr_tokens-2,tokens+2);
-		}
+		for(int i=0;i<forcount;i++) {run_command(nr_tokens-2,tokens+2);}
 		return 1;
 	} // "for" is built in
 
 	else if(!strcmp(tokens[0],"prompt")){
 		strcpy(__prompt,tokens[1]);
-		return 1; // 여기서 끝
+		return 1;
 	} // "prompt" is built in
 
 	else if (!strcmp(tokens[0],"timeout")){
-
-			if(nr_tokens==1) {
-				printf("Current timeout is %d second\n",__timeout);
-				return 1;
-			}
-			else{
-				set_timeout(atoi(tokens[1]));
-				//alarm(__timeout);
-				//__timeout 이 시간 값
-				// alarm 시그널 생기면 sa에 등록된 핸들러 실행
-				return 1;
-			}
-		// 여기서 끝
+		if(nr_tokens==1) {
+			printf("Current timeout is %d second\n",__timeout);
+			return 1;
 		}
-
-
+		else{
+			set_timeout(atoi(tokens[1]));
+			return 1;
+		}
+	}
 /********************************************************************/
 	// 자식 프로세스 생성
 	if((pidnum=fork())==0) {
 		alarm(__timeout); // 자식프로세스 만들자마자 alarm 대기
-		sigaction(SIGALRM,&sa,NULL); // alarm 신호 생기면 바로 action 취한다.
-		if(execvp(tokens[0],tokens)<0){
-			fprintf(stderr,"No such file or directory\n");
-			// confirm no such file
-			// execl("/bin/echo","/bin/echo","No such file or directory",NULL);
-			// 이 명령어로 해당 문장 출력하면서 죽으면 정상적으로 죽지만 
-
-			kill(getpid(),SIGKILL);
-			// 여기서 exit 로 프로세스를 죽이면 자식 프로세스의 출력 버퍼를 던지고 죽는다.
-		}
+		if(execvp(tokens[0],tokens)<0) kill(getpid(),SIGKILL);	// 잘못 입력시
 		else return 1;
 	}
-	 // child
- 	
+
 	else if(pidnum==-1){
-		wait(&status);
+		printf("fork error\n");
 		return 0;
 	} // error
 
 	else { 
-		// WIFSIGNALERD(status) 이 걸로 자식이 어떤 신호를 받아 죽
-		// 었으면 바로 알 수 있지만 어떤 시그널인지
-		// 확인 할 수 없기 때문에 WTERMSIG 를 쓴다.
 		if((waitPid=wait(&status))!=-1) // child is closed?
 		{
 			if(WTERMSIG(status)==SIGALRM){ // 프로세스가 알람신호로 죽엇나?
@@ -181,15 +133,13 @@ static int run_command(int nr_tokens, char *tokens[])
 			}
 			else if(WTERMSIG(status)==SIGKILL) // 프로세스가 kill 신호로 죽엇나?
 			{
+				fprintf(stderr,"No such file or directory\n");
 				return 1;
 			}
 			else return 1;
 		}
-		else return 0;
+		else return 0; // wait 함수 오류 ( wait 의 반환값 = -1 )
 	} // parent
-
-	
-	return 1;
 }
 
 
@@ -253,7 +203,7 @@ int main(int argc, char * const argv[])
 	if ((ret = initialize(argc, argv))) return EXIT_FAILURE;
 
 	if (__verbose)
-		fprintf(stderr, "%s%s%s ", __color_start, __prompt, __color_end);
+		fprintf(stderr, "%s%s%s%s ", __color_start, "FuckU",__prompt, __color_end);
 
 	while (fgets(command, sizeof(command), stdin)) {	
 		char *tokens[MAX_NR_TOKENS] = { NULL };
@@ -272,7 +222,7 @@ int main(int argc, char * const argv[])
 
 more:
 		if (__verbose)
-			fprintf(stderr, "%s%s%s ", __color_start, __prompt, __color_end);
+			fprintf(stderr, "%s%s%s%s ", __color_start,"FuckU", __prompt, __color_end);
 	}
 
 	finalize(argc, argv);
